@@ -1,7 +1,7 @@
 //  Copyright (c) 2018 Loup Inc.
 //  Licensed under Apache License v2.0
 
-part of beacons;
+part of omnys_beacons;
 
 class BeaconRegionRequest {
   BeaconRegionRequest({this.android, this.iOS}) {
@@ -10,9 +10,9 @@ class BeaconRegionRequest {
     // do not allow different identifier for each platform
     // it does not make sense and could result in subtle bugs for the end user
     assert((android == null || iOS == null) ||
-        android.identifier == iOS.identifier);
+        android?.identifier == iOS?.identifier);
   }
 
-  final BeaconRegion android;
-  final BeaconRegionIBeacon iOS;
+  final BeaconRegion? android;
+  final BeaconRegionIBeacon? iOS;
 }
